+++ src/Controller/ManagementReportController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Form\ManagementReportDateIntervalType;
 use App\Repository\InvoiceRepository;
+use App\Service\ManagementReportService;
 use App\Service\ViewService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -30,7 +31,10 @@ class ManagementReportController extends AbstractController
 
         $form = $this->createForm(
             ManagementReportDateIntervalType::class,
-            ['firstLog' => $firstRecordedInvoice->getRecordedDate()],
+            [
+                'firstLog' => $firstRecordedInvoice->getRecordedDate(),
+                'view' => $this->viewService->getCurrentViewId(),
+            ],
             ['action' => $this->generateUrl('app_management_reports_output', $this->viewService->addView([])), 'method' => 'GET']
         );
 
@@ -46,16 +50,48 @@ class ManagementReportController extends AbstractController
     public function output(Request $request, InvoiceRepository $invoiceRepository): Response
     {
         $queryElements = $request->query->all();
+        $view = $queryElements['view'];
         $dateInterval = $queryElements['management_report_date_interval'] ?? null;
 
         if (empty($dateInterval['dateFrom']) || empty($dateInterval['dateTo'])) {
             return $this->redirectToRoute('app_management_reports_create', $this->viewService->addView([]), Response::HTTP_SEE_OTHER);
         }
-        $invoices = $invoiceRepository->getByRecordedDateBetween(
-            new \DateTime($dateInterval['dateFrom']),
-            new \DateTime($dateInterval['dateTo'].' 23:59:59')
-        );
 
+        $invoices = $this->getInvoicesDataFromDates($dateInterval, $invoiceRepository, $view);
+
+        return $this->render(
+            'management-report/output.html.twig',
+            [
+                'groupedInvoices' => $this->createGroupedInvoices($invoices),
+                'dateInterval' => $dateInterval,
+                'view' => $queryElements['view'],
+                'currentQuery' => $queryElements,
+            ]
+        );
+    }
+
+    /**
+     * @throws \Doctrine\ORM\Exception\NotSupported
+     * @throws \Exception
+     */
+    #[Route('/output/export', name: 'app_management_reports_output_export', methods: ['GET'])]
+    public function export(Request $request, ManagementReportService $managementReportService, InvoiceRepository $invoiceRepository): Response
+    {
+        $queryElements = $request->query->all();
+        $view = $queryElements['view'];
+        $dateInterval = $queryElements['management_report_date_interval'] ?? null;
+
+        if (empty($dateInterval['dateFrom']) || empty($dateInterval['dateTo'])) {
+            return $this->redirectToRoute('app_management_reports_create', [], Response::HTTP_SEE_OTHER);
+        }
+
+        $invoices = $this->getInvoicesDataFromDates($dateInterval, $invoiceRepository, $view);
+
+        return $managementReportService->generateSpreadsheetCsvResponse($this->createGroupedInvoices($invoices), $dateInterval);
+    }
+
+    private function createGroupedInvoices(array $invoices): array
+    {
         $groupedInvoices = [];
         foreach ($invoices as $invoice) {
             $recordedDate = $invoice->getRecordedDate();
@@ -85,12 +121,19 @@ class ManagementReportController extends AbstractController
             $groupedInvoices[$year] = array_merge(['sum' => $sum], $groupedInvoices[$year]);
         }
 
-        return $this->render(
-            'management-report/output.html.twig',
-            $this->viewService->addView([
-                'groupedInvoices' => $groupedInvoices,
-                'dateInterval' => $dateInterval,
-            ])
+        return $groupedInvoices;
+    }
+
+    /**
+     * @throws \Doctrine\ORM\Exception\NotSupported
+     * @throws \Exception
+     */
+    private function getInvoicesDataFromDates($dateInterval, InvoiceRepository $invoiceRepository, ?string $viewId): array
+    {
+        return $invoiceRepository->getByRecordedDateBetween(
+            new \DateTime($dateInterval['dateFrom']),
+            new \DateTime($dateInterval['dateTo'].' 23:59:59'),
+            $viewId
         );
     }
 }