+++ src/Entity/Project.php
@@ -49,6 +49,12 @@ class Project extends AbstractBaseEntity
     #[ORM\ManyToMany(targetEntity: View::class, mappedBy: 'projects')]
     private Collection $views;
 
+    #[ORM\Column(length: 255, nullable: true)]
+    private ?string $projectLeadName = null;
+
+    #[ORM\Column(length: 255, nullable: true)]
+    private ?string $projectLeadMail = null;
+
     public function __construct()
     {
         $this->invoices = new ArrayCollection();
@@ -313,6 +319,30 @@ class Project extends AbstractBaseEntity
         if ($this->views->removeElement($view)) {
             $view->removeProject($this);
         }
+
+        return $this;
+    }
+
+    public function getProjectLeadName(): ?string
+    {
+        return $this->projectLeadName;
+    }
+
+    public function setProjectLeadName(?string $projectLeadName): static
+    {
+        $this->projectLeadName = $projectLeadName;
+
+        return $this;
+    }
+
+    public function getProjectLeadMail(): ?string
+    {
+        return $this->projectLeadMail;
+    }
+
+    public function setProjectLeadMail(?string $projectLeadMail): static
+    {
+        $this->projectLeadMail = $projectLeadMail;
 
         return $this;
     }