+++ src/Model/Invoices/AccountData.php
@@ -7,15 +7,11 @@ class AccountData
     public readonly string $projectTrackerId;
     public readonly string $name;
     public readonly string $value;
-    public readonly ?string $category;
-    public readonly ?string $status;
 
-    public function __construct(string $projectTrackerId, string $name, string $value, ?string $category, ?string $status)
+    public function __construct(string $projectTrackerId, string $name, string $value)
     {
         $this->projectTrackerId = $projectTrackerId;
         $this->name = $name;
         $this->value = $value;
-        $this->category = $category;
-        $this->status = $status;
     }
 }